backend/requirements.txt
# Core
fastapi==0.68.2
uvicorn==0.15.0
pydantic==1.9.0
aiohttp==3.8.1

# Data Processing
geopandas==0.10.2
pyarrow==6.0.1

# Storage
google-cloud-storage==1.43.0

# Development
python-dotenv==0.19.2
pytest==6.2.5
pytest-asyncio==0.16.0
